Fix getting workflow access context for versions
